Move test edition to test fixtures
