+++ app/editions/EditionTemplates.scala
@@ -9,7 +9,6 @@ object WeekDay extends Enumeration(0) {
   val Mon, Tues, Wed, Thurs, Fri, Sat, Sun = Value
   implicit def WeekDayToInt(weekDay: WeekDay): Int = weekDay.id
   implicit def IntToWeekDay(int: Int): WeekDay = WeekDay(int)
-  //def LocalDateToWeekday: WeekDay = ???
 }
 
 case class FrontPresentation()
@@ -52,47 +51,6 @@ case class EditionTemplateForDate(
   name: String,
   fronts: List[FrontTemplate],
 )
-
-
-// ??? Is this how we want to model collections??
-object Sport {
-  val emptyCollectionPresentation = CollectionPresentation()
-  val football: CollectionTemplate = new CollectionTemplate("Football", CapiQuery("???"), emptyCollectionPresentation)
-  val cricket: CollectionTemplate = new CollectionTemplate("Cricket", CapiQuery("???"), emptyCollectionPresentation)
-}
-
-object Fronts {
-  val emptyFrontPresentation = FrontPresentation()
-  val ukNews: FrontTemplate = FrontTemplate("UK news", List(), emptyFrontPresentation)
-  val sports: FrontTemplate = FrontTemplate(
-    "Sport",
-    List(
-      Sport.football,
-      Sport.cricket,
-    ),
-    emptyFrontPresentation
-  )
-  val technology: FrontTemplate = FrontTemplate(
-    "Technology",
-    List(),
-    emptyFrontPresentation
-  )
-
-  val opinion: FrontTemplate = FrontTemplate("Opinion", List(), emptyFrontPresentation)
-}
-
-object DailyEdition {
-  val dailyEdition: EditionTemplate = EditionTemplate(
-    "dailyEdition",
-    List(
-      (Fronts.ukNews, Daily()),
-      (Fronts.sports, Daily()),
-      (Fronts.opinion, WeekDays(List(Thurs, Wed))),
-      (Fronts.technology, WeekDays(List(Thurs)))
-    ),
-    Daily()
-  )
-}
 
 object EditionTemplateHelpers {
   def generateEditionTemplate(date: DateTime, edition: EditionTemplate): Option[EditionTemplateForDate] = {
